First commit with http POST json to backend api
As a reference the json document for PV-Update used by
map-pannel-component.ts in the angular front-end has been used here.
This worked fine and now a new json document will be made to be used as
lab-update in the backend api.
Back-end api will also need to be changed for that purpose